requirements.txt
crewai>=0.130.0
openai>=1.30.0
python-dotenv>=1.0.1